initial changes to get the reject task converted, just need to clean up the test
