+++ mail/tasks.py
@@ -2,8 +2,6 @@ import logging
 import os
 import urllib.parse
 from datetime import timedelta
-from email.mime.multipart import MIMEMultipart
-from email.mime.text import MIMEText
 from typing import List, MutableMapping, Tuple
 
 from background_task import background
@@ -21,7 +19,6 @@ from mail.libraries.lite_to_edifact_converter import EdifactValidationError
 from mail.libraries.routing_controller import check_and_route_emails, send, update_mail
 from mail.libraries.usage_data_decomposition import build_json_payload_from_data_blocks, split_edi_data_by_id
 from mail.models import LicenceIdMapping, LicencePayload, Mail, UsageData
-from mail.servers import smtp_send
 
 logger = logging.getLogger(__name__)
 
@@ -252,38 +249,7 @@ def send_licence_data_to_hmrc():
         return True
 
 
-# Notify Users of Rejected Mail
-@background(queue=NOTIFY_USERS_TASK_QUEUE, schedule=0)
-def notify_users_of_rejected_mail(mail_id, mail_response_date):
-    """If a rejected email is found, this task notifies users of the rejection"""
-
-    logger.info("Notifying users of rejected Mail [%s, %s]", mail_id, mail_response_date)
-
-    try:
-        multipart_msg = MIMEMultipart()
-        multipart_msg["From"] = settings.EMAIL_USER
-        multipart_msg["To"] = ",".join(settings.NOTIFY_USERS)
-        multipart_msg["Subject"] = "Mail rejected"
-        body = MIMEText(f"Mail [{mail_id}] received at [{mail_response_date}] was rejected")
-        multipart_msg.attach(body)
-
-        smtp_send(multipart_msg)
-    except Exception as exc:  # noqa
-        error_message = (
-            f"An unexpected error occurred when notifying users of rejected Mail "
-            f"[{mail_id}, {mail_response_date}] -> {type(exc).__name__}: {exc}"
-        )
-
-        # Raise an exception
-        # this will cause the task to be marked as 'Failed' and retried if there are retry attempts left
-        raise Exception(error_message)
-    else:
-        logger.info("Successfully notified users of rejected Mail [%s, %s]", mail_id, mail_response_date)
-
-
 # Manage Inbox
-
-
 @background(queue=MANAGE_INBOX_TASK_QUEUE, schedule=0)
 def manage_inbox():
     """Main task which scans inbox for SPIRE and HMRC emails"""
